requirements.txt
numpy
Pillow==6.2.1
tensorflow==1.14.0
tqdm